+++ mkdocs.yml
@@ -19,6 +19,9 @@ theme:
 # an error when building the docs.
 markdown_extensions:
   - footnotes
+  - pymdownx.superfences
+  - pymdownx.details # add notes toggleable notes ???
+
 
 plugins:
   - search
@@ -51,6 +54,7 @@ plugins:
 
 extra_javascript:
   - javascripts/katex.js
+  - javascripts/toggle_code.js
   - https://cdnjs.cloudflare.com/ajax/libs/KaTeX/0.16.7/katex.min.js
   - https://cdnjs.cloudflare.com/ajax/libs/KaTeX/0.16.7/contrib/auto-render.min.js
